refactor: common 모듈 quiz, study 관련 모델 패키지 분리
