--- common/src/main/java/com/paykidscompose/common/model/ChatResponseModel.kt
@@ -1,5 +1,0 @@
-package com.paykidscompose.common.model
-
-data class ChatResponseModel (
-    val response: Map<String, String>
-): Model()
