+++ common/src/main/java/com/paykidscompose/common/model/quiz/WrongAnswerNoteStatus.kt
@@ -1,4 +1,4 @@
-package com.paykidscompose.common.model
+package com.paykidscompose.common.model.quiz
 
 sealed class WrongAnswerNoteStatus {
     object NoAttempt : WrongAnswerNoteStatus()
