+++ common/src/main/java/com/paykidscompose/common/model/study/ChatResponseModel.kt
@@ -1,0 +1,7 @@
+package com.paykidscompose.common.model.study
+
+import com.paykidscompose.common.model.Model
+
+data class ChatResponseModel (
+    val response: Map<String, String>
+): Model()
