+++ common/src/main/java/com/paykidscompose/common/usecase/study/GetChatResponseUseCase.kt
@@ -1,7 +1,7 @@
 package com.paykidscompose.common.usecase.study
 
 import com.paykidscompose.common.exception.PayKidsException
-import com.paykidscompose.common.model.ChatResponseModel
+import com.paykidscompose.common.model.study.ChatResponseModel
 import com.paykidscompose.common.repositories.ChatRepository
 import com.paykidscompose.common.result.DataResourceResult
 import com.paykidscompose.common.usecase.base.FlowUseCase
