+++ data/src/main/java/com/paykidscompose/data/mapper/quiz/QuizClearedMapper.kt
@@ -1,7 +1,7 @@
 package com.paykidscompose.data.mapper.quiz
 
 import com.paykidscompose.common.mapper.ModelMapper
-import com.paykidscompose.common.model.QuizClearedModel
+import com.paykidscompose.common.model.quiz.QuizClearedModel
 import com.paykidscompose.data.model.quiz.QuizClearedDTO
 
 object QuizClearedMapper : ModelMapper<QuizClearedModel, QuizClearedDTO> {
