+++ data/src/main/java/com/paykidscompose/data/repositories/ChatRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.paykidscompose.data.repositories
 
 import com.paykidscompose.common.exception.PayKidsException
-import com.paykidscompose.common.model.ChatResponseModel
+import com.paykidscompose.common.model.study.ChatResponseModel
 import com.paykidscompose.common.repositories.ChatRepository
 import com.paykidscompose.common.result.DataResourceResult
 import com.paykidscompose.data.mapper.study.ChatResponseMapper
