+++ data/src/main/java/com/paykidscompose/data/repositories/QuizRepositoryImpl.kt
@@ -1,8 +1,8 @@
 package com.paykidscompose.data.repositories
 
 import com.paykidscompose.common.exception.PayKidsException
-import com.paykidscompose.common.model.QuizClearedModel
-import com.paykidscompose.common.model.QuizModel
+import com.paykidscompose.common.model.quiz.QuizClearedModel
+import com.paykidscompose.common.model.quiz.QuizModel
 import com.paykidscompose.common.repositories.QuizRepository
 import com.paykidscompose.common.result.DataResourceResult
 import com.paykidscompose.data.mapper.quiz.QuizClearedMapper
